+++ week3/yourChoice/src/components/Prefer/Prefer.jsx
@@ -1,42 +1,12 @@
 import { useState } from 'react';
 import styled from 'styled-components';
 
+import { CONTENT_HEADER, SONG_DATA, SONG_TITLE } from '../../assets/song_data/songData';
 import ContentWrapper from '../UI/ContentWrapper';
 import ContentHeader from '../UI/ContentHeader';
 import H1 from '../UI/H1';
 
 const Prefer = (props) => {
-  const CONTENT_HEADER = [
-    '지금 무슨 노래를 듣고싶어?',
-    '그럼 이 중에서는 뭐가 끌려?',
-    '마지막으로 선택해 줘!',
-    '오늘의 추천 플리는 바로!',
-  ];
-
-  const SONG_DATA = [
-    ['인디/밴드', 'k-pop', '팝송'],
-    ['신나는', '잔잔한', '주인장 추천'],
-    ['여자', '남자', '상관없어!'],
-  ];
-
-  const SONG_TITLE = [
-    [
-      ['제이레빗 happy things', '데이식스 한 페이지가 될 수 있게', 'LUCY 아니근데진짜'],
-      ['최유리 바람', '데이식스 hihello', '데이먼스이어 yours'],
-      ['옥상달빛 수고했어 오늘도', '하현상 등대', 'KozyPop Dezavu'],
-    ],
-    [
-      ['아이브 iam', '부석순 파이팅해야지', '전소미 fast forward'],
-      ['뉴진스 Ditto', 'EXO 첫눈', '수지 cape'],
-      ['STAYC so what', '엔시티 드림 미니카', '선미 보랏빛 밤'],
-    ],
-    [
-      ['Taylor Swift Shake it off', 'Sam Smith Unholy', 'Justin Bieber, 라로이 STAY'],
-      ['Selena Gomez lose you to love me', 'Justin Bieber off my face', '5S0S high'],
-      ['Selena Gomez Who says', "Charlie Puth That's hilarious", "Why don't we What am I"],
-    ],
-  ];
-
   // 단계 확인용
   const [step, setStep] = useState(0);
